core: schema v5 (int16 battery, mag-rate heading); app Variables + coarse real series
Translation localization: drive the Kalman rate from the magnetometer's own rotation
rate (spin_rate_lp) while the mag is valid, instead of the accel magnitude. Linear
body acceleration during a drift adds a once-per-rev term to the accel that both
wobbles the rate (integrating into ~30 deg heading swings — the "LED fluctuates")
and biases it high (Jensen -> slow drift); the B-field's rotation has neither
problem. Replay on real logs: translation LED excursion 33->23 deg (toward the ~19
deg pure-spin baseline), drift down to the pure-spin level, no penalty during rapid
spin changes, pure spin unchanged. The band-pass centre is now LP-smoothed
(spin_freq_lp) so a time-varying filter no longer injects heading wobble; it stays
open-loop off the accel (never kf_omega). New bringup knobs MAG_BP_FC_LP_HZ and
MAG_SPIN_RATE_LP_HZ.

Schema v5:
- batt_offset int8 -> int16 (1 mV/LSB, was 20.5 mV): telemetry no longer discretises
  the battery (ADC ~2.4 mV + 6 Hz LP are the only quantisers left). Log readers are
  schema-aware (replay.rs / replay.c remap the pre-v5 int8), so all existing logs
  still parse (verified: mode counts bit-identical). Battery scale updated everywhere
  it is decoded (simulation, StatusBar/TopBar/DriverStation).
- SunshineState grows by spin_freq_lp. Frame 687 -> 715 B; the receiver's
  ESPNOW_TELEM_SIZE is bumped to match (it drops wrong-length frames -> would have
  been total telemetry loss). Wire format changed: reflash brain + receiver together.

Scope app:
- New Variables channel group (battery, eRPM, omega-from-accel, centripetal, accel
  m/s^2): input-only quantities that are identical for real and replayed, so one
  shared 1 kHz series instead of a redundant real/replayed split (and off the Inputs
  group for accel m/s^2). Each unit gets its own axis.
- Real series is emitted only at the true 100 Hz snapshots and drawn as coarse lines
  (frontend union-x + spanGaps), not a fabricated 1 kHz staircase-with-noise that
  matched the replayed dot count.

Docs: ARCHITECTURE/DEBUGGING schema + sizes; TUNING/BRINGUP mag-rate heading,
band-pass LP, RPM-vs-translation bandwidth + data-derived phase-lead, inverted-frame
direction note.

+++ sunshine_core/src/utils.c
@@ -36,5 +36,5 @@ uint16_t sunshine_f32_to_f16(float val) {
 
 float sunshine_accel_to_ms2(int16_t raw) { return (float)raw * ADXL_SCALE_MS2; }
 float sunshine_mag_to_ut   (int16_t raw) { return (float)raw * MAG_SCALE_UT;   }
-float sunshine_batt_to_v   (int8_t  off) { return BATT_OFFSET_REF_V + (float)off * BATT_SCALE_V; }
+float sunshine_batt_to_v   (int16_t off) { return BATT_OFFSET_REF_V + (float)off * BATT_SCALE_V; }
 uint32_t sunshine_schema_version(void)   { return SUNSHINE_SCHEMA_VERSION; }
